% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOncoplotGSEA.R
\name{GetOncoplotGSEA}
\alias{GetOncoplotGSEA}
\title{GetOncoplotGSEA}
\usage{
GetOncoplotGSEA(
  tbl_overlap,
  tbl_spl = NULL,
  group_colors = c("tomato1", "limegreen", "blue"),
  group_labels = NULL,
  name = "p<0.05",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
GetOncoplotGSEA
}
