% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOncoplotGSEA_Bi.R
\name{GetOncoplotGSEA_Bi}
\alias{GetOncoplotGSEA_Bi}
\title{GetOncoplotGSEA_Bi}
\usage{
GetOncoplotGSEA_Bi(
  tbl_overlap,
  tbl_spl = NULL,
  group_colors = c("tomato1", "limegreen", "blue"),
  group_labels = NULL,
  name = "p<0.05",
  visual_query_as_row = F,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
GetOncoplotGSEA_Bi
}
