% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happyORA.R
\name{happyORA}
\alias{happyORA}
\title{happyORA}
\usage{
happyORA(
  query_input,
  colors = c("red", "blue", "limegreen"),
  remove_genesets_group = NULL,
  keep_genesets_group = NULL,
  background = NULL,
  p_cut = 0.05
)
}
\arguments{
\item{query_input}{a two-column table with column name as "cluster" and "gene_id"}

\item{remove_genesets_group}{a vector or a string. group names of gene sets not to show
"Amplicon", "Cancer_related", "CHIP", "DNA_Damage_Repair", "Epigenetic",
"Gene Targets", "Hallmark", "Other Targets", "Pathway Targets"}

\item{keep_genesets_group}{select a group of gene sets only}

\item{background}{if null, genome-wide genes are used}

\item{p_cut}{default 0.05}
}
\description{
happyORA
}
