% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleGSEA.R
\name{SimpleGSEA}
\alias{SimpleGSEA}
\title{SimpleGSEA}
\usage{
SimpleGSEA(
  query_input,
  genesets_list,
  background = NULL,
  visual = T,
  p_cut = 0.05,
  cols = NULL
)
}
\arguments{
\item{query_input}{two column tibble with "gene_id" and "cluster" as group}

\item{genesets_list}{lists of gene sets}

\item{background}{can be genome-wide. if not provide,
bg will be total genes of genesets_list}

\item{p_cut}{0.05 by default}

\item{cols}{colors in order}
}
\description{
SimpleGSEA
}
